Separated out code for applying a search query to a set of items, and code for event binding and showing results
